Add authorization service

+++ Restaurants.Application/Dishes/Commands/DeleteAllDishesForRestaurant/DeleteAllDishesForRestaurantCommandHandler.cs
@@ -1,5 +1,7 @@
 using MediatR;
 using Microsoft.Extensions.Logging;
+using Restaurants.Domain.Constants;
+using Restaurants.Domain.Contracts;
 using Restaurants.Domain.Entities;
 using Restaurants.Domain.Exceptions;
 using Restaurants.Domain.Repositories;
@@ -9,13 +11,21 @@ namespace Restaurants.Application.Dishes.Commands.DeleteAllDishesForRestaurant;
 public class DeleteAllDishesForRestaurantCommandHandler(
     ILogger<DeleteAllDishesForRestaurantCommandHandler> logger,
     IRestaurantsRepository restaurantsRepository,
-    IDishesRepository dishesRepository) : IRequestHandler<DeleteAllDishesForRestaurantCommand>
+    IDishesRepository dishesRepository,
+    IRestaurantAuthorizationService restaurantAuthorizationService)
+    : IRequestHandler<DeleteAllDishesForRestaurantCommand>
+
 {
     public async Task Handle(DeleteAllDishesForRestaurantCommand request, CancellationToken cancellationToken)
     {
         logger.LogInformation("Deleting all dishes for restaurant : {RestaurantId}", request.RestaurantId);
         var restaurant = await restaurantsRepository.GetById(request.RestaurantId);
         if (restaurant == null) throw new NotFoundException(nameof(Restaurant), request.RestaurantId.ToString());
+        if (!restaurantAuthorizationService.Authorize(restaurant, ResourceOperation.Update))
+        {
+            throw new ForbidException();
+        }
+
         await dishesRepository.Delete(restaurant.Dishes);
     }
 }
